split oracles conversion into file
